split GamepadSpeakerApp into separate module
