+++ README.md
@@ -56,7 +56,7 @@
 天気予報は気象庁から取得します。ボタンに割り当てる処理を変更したい場合は、`main.py` を編集するか、次のようなスクリプトを作成してください。
 
 ```python
-from main import GamepadSpeakerApp
+from gamepad_speaker_app import GamepadSpeakerApp
 from tasks.time import fetch_time
 from tasks.weather import fetch_weather_today
